// 02_REFS_PROTOTYPES/XPRT/packages/xprt-firecrawl/firecrawl-searxng/apps/api/src/ai/types.ts
export interface Configuration {
  activeAIProvider: string;
  openRouterApiKey?: string;
  openRouterModel?: string;
  geminiApiKey?: string;
  mistralApiKey?: string;
  requestyApiKey?: string;
  groqApiKey?: string;
  cerebrasApiKey?: string;
  githubModelsApiKey?: string;
  githubModelsEndpoint?: string;
  huggingfaceApiKey?: string;
  huggingfaceModel?: string;
}